+++ packages/ndla-ui/src/FactBox/FactBox.tsx
@@ -6,10 +6,12 @@
  *
  */
 
-import { ReactNode, MouseEvent } from 'react';
+import { ReactNode, useState } from 'react';
 import BEMHelper from 'react-bem-helper';
-import { ButtonV2 } from '@ndla/button';
 import { useTranslation } from 'react-i18next';
+import styled from '@emotion/styled';
+import { IconButtonV2 } from '@ndla/button';
+import { ChevronDown, ChevronUp } from '@ndla/icons/common';
 
 const classes = new BEMHelper({
   name: 'factbox',
@@ -17,26 +19,37 @@ const classes = new BEMHelper({
 });
 
 interface Props {
-  dangerouslySetInnerHTML?: { __html: string };
   children?: ReactNode;
 }
 
-const toggleFactBox = (event: MouseEvent<HTMLButtonElement>) => {
-  const button = event.currentTarget;
-  const aside = button?.previousSibling?.parentElement;
-  aside?.classList?.toggle('expanded');
-};
-const FactBox = ({ children, dangerouslySetInnerHTML }: Props) => {
+const StyledAside = styled.aside`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+const StyledDiv = styled.div`
+  width: 100%;
+`;
+
+const StyledIconButton = styled(IconButtonV2)`
+  margin-top: -20px;
+  z-index: 1;
+`;
+
+const FactBox = ({ children }: Props) => {
   const { t } = useTranslation();
+  const [isOpen, setIsOpen] = useState(false);
+
+  const additional = isOpen ? 'expanded' : '';
 
   return (
-    <aside {...classes()}>
-      <div {...classes('content')} dangerouslySetInnerHTML={dangerouslySetInnerHTML}>
-        {children}
-      </div>
-      <ButtonV2 {...classes('button', 'collapsed')} onClick={toggleFactBox} title={t('factbox.open')} />
-      <ButtonV2 {...classes('button', 'open')} onClick={toggleFactBox} title={t('factbox.close')} />
-    </aside>
+    <StyledAside {...classes(undefined, undefined, additional)}>
+      <StyledDiv {...classes('content')}>{children}</StyledDiv>
+      <StyledIconButton onClick={() => setIsOpen((p) => !p)} aria-label={t(`factbox.${isOpen ? 'close' : 'open'}`)}>
+        {isOpen ? <ChevronUp /> : <ChevronDown />}
+      </StyledIconButton>
+    </StyledAside>
   );
 };
