+++ packages/ndla-ui/src/FactBox/index.ts
@@ -1,5 +1,3 @@
 import FactBox from './FactBox';
 
-export { default as FactBoxV2 } from './FactBoxV2';
-
 export default FactBox;
